[HaCreator+HaRepacker] Move WzFileManager to MapleLib proj
-- needed for the next phase (fix '_outlink' property for 64-bit wz file format)
